requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.1
sqlalchemy>=2.0.0
pymysql>=1.1.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
email-validator>=2.1.0
